Add files via upload
ICT Olympiad Bangladesh (ICTOB) is one of the largest and most prominent yearly ICT Olympiad events in Bangladesh. It is held for students of all educational levels, from pre-school to university. Every student in Bangladesh is eligible to participate in the competition. This is a major event in the field of information and communications technology, and it covers a broad spectrum of topics, such as ICT Fundamentals, Communication System and Networking, Number System and Digital Devices, Web Design and HTML, Basic Programming, and Database Management System, among many others. Students who are studying for the Digital Bangladesh exam will find that using the suggestions offered in this article helps them feel more prepared and confident. The participants will gain a deeper understanding of new ideas and concepts, which will not only help them improve their self-confidence in the realm of information and communications technology (ICT), but will also enable them to better shape their futures within this dynamic industry by making effective use of the ICT skills they have acquired. This is because the participants will gain a deeper understanding of new ideas and concepts, which will help them improve their self-confidence in the realm of ICT. The Information and Communication Technology Olympiad Bangladesh (ICTOB) will be an incredible trip with the subject "Road to New Revolution," and it will feature participation from a considerable number of pupils. The Creative It Institute and Whistle are in the process of taking their olympiad competition on the road to several locations across the United States. Individuals that are strong leaders and believe strongly in what they stand for make up a Campus Ambassador team that is extremely dedicated to their work and very active. Because "Road to a New Revolution" is our motto, we are putting in a lot of effort to ensure that our platform is capable of housing the most advanced, world-class technologically adaptive facilities possible. In this location, we are providing training for ambassadors to prepare them to meet the difficulties of the post-pandemic era with conviction and compassion. ICT Olympiad Bangladesh will soon take the initiative to expand the ICT Olympiad Bangladesh to an international level. This will make it possible for students from a broad variety of nations to compete against one another in the same event. On this stage, students from all over the world would have the opportunity to demonstrate their skills in the field of information and communications technology, and students from Bangladesh would have the chance to compete on an international level. Students will be able to finish their coursework and see the most recent video content within this app. In this section, students are also able to apply to become campus ambassadors.